Corrigiendo Errores

+++ src/main/java/com/edu/mx/lasalle/oaxaca/servicio_aeropuerto/service/implement/BoletoServiceImplements.java
@@ -8,7 +8,11 @@ import com.edu.mx.lasalle.oaxaca.servicio_aeropuerto.models.BoletoModel;
 import com.edu.mx.lasalle.oaxaca.servicio_aeropuerto.repositories.BoletoRepository;
 import com.edu.mx.lasalle.oaxaca.servicio_aeropuerto.service.BoletoService;
 import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  *
@@ -17,35 +21,56 @@ import org.springframework.beans.factory.annotation.Autowired;
 public class BoletoServiceImplements implements BoletoService {
     @Autowired
     private BoletoRepository boletoRepository;
+
+    @Transactional
     @Override
-    public void registrarBoleto(BoletoModel boletoModel) {
-        boletoRepository.save(boletoModel);
+    public BoletoModel registrarBoleto(BoletoModel boletoModel) {
+        return boletoRepository.save(boletoModel);
     }
 
+    @Transactional(readOnly = true)
     @Override
     public List<BoletoModel> obtenerBoletos() {
         return (List<BoletoModel>) boletoRepository.findAll();
     }
 
+    @Transactional(readOnly = true)
     @Override
-    public BoletoModel getBoleto(int id) {
-        return boletoRepository.findById(id);
+    public Optional<BoletoModel> getBoleto(int id) {
+        return boletoRepository.findByIdBoleto(id);
     }
 
+    @Transactional
+    @Modifying
     @Override
-    public void actualizarDatosBoleto(BoletoModel boletoModel, int id) {
-        boletoModel.setIdBoleto(id);
-        boletoRepository.save(boletoModel);
+    public Optional<BoletoModel> actualizarDatosBoleto(BoletoModel boletoModel, int id) {
+        Optional<BoletoModel> boletoOptional = boletoRepository.findById(id);
+        if (boletoOptional.isPresent()) {
+            BoletoModel boletoActualizado = boletoOptional.orElseThrow();
+            boletoActualizado.setAsiento(boletoModel.getAsiento());
+            boletoActualizado.setCosto(boletoModel.getCosto());
+            boletoActualizado.setVueloModel(boletoModel.getVueloModel());
+            return Optional.of(boletoRepository.save(boletoActualizado));
+        }
+        return Optional.empty();
     }
 
+    @Transactional
+    @Modifying
     @Override
-    public void borrarBoleto(int id) {
-        boletoRepository.deleteById(id);
+    public Optional<BoletoModel> borrarBoleto(int id) {
+        Optional<BoletoModel> boletoOptional = boletoRepository.findById(id);
+        if (boletoOptional.isPresent()) {
+            boletoRepository.deleteById(id);
+            return boletoOptional;
+        }
+        return Optional.empty();
     }
 
+    @Transactional
+    @Modifying
     @Override
     public void borrarTodosLosBoletos() {
         boletoRepository.deleteAll();
     }
-    
 }
